got allItems working alone
